use zod for server actions and form, add validations.ts for zod schemas

+++ src/actions/actions.ts
@@ -2,12 +2,19 @@
 
 import prisma from "@/lib/db";
 import { TCar } from "@/lib/types";
+import { carFormSchema, carIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function addCar(car: Omit<TCar, "id">) {
+export async function addCar(car: unknown) {
+  const validatedCar = carFormSchema.safeParse(car);
+  if (!validatedCar.success) {
+    return {
+      message: "Invalid data. Please check your inputs and try again.",
+    };
+  }
   try {
     await prisma.car.create({
-      data: car,
+      data: validatedCar.data,
     });
   } catch (error) {
     return {
@@ -17,11 +24,18 @@ export async function addCar(car: Omit<TCar, "id">) {
   revalidatePath("/valet", "layout");
 }
 
-export async function editCar(carId: string, car: Omit<TCar, "id">) {
+export async function editCar(carId: unknown, car: unknown) {
+  const validatedCarId = carIdSchema.safeParse(carId);
+  const validatedCar = carFormSchema.safeParse(car);
+  if (!validatedCar.success || !validatedCarId.success) {
+    return {
+      message: "Invalid data. Please check your inputs and try again.",
+    };
+  }
   try {
     await prisma.car.update({
-      where: { id: carId },
-      data: car,
+      where: { id: validatedCarId.data },
+      data: validatedCar.data,
     });
   } catch (error) {
     return {
@@ -31,10 +45,16 @@ export async function editCar(carId: string, car: Omit<TCar, "id">) {
   revalidatePath("/valet", "layout");
 }
 
-export async function checkOutCar(carId: string) {
+export async function checkOutCar(carId: unknown) {
+  const validatedCarId = carIdSchema.safeParse(carId);
+  if (!validatedCarId.success) {
+    return {
+      message: "Invalid data. Please check your inputs and try again.",
+    };
+  }
   try {
     await prisma.car.delete({
-      where: { id: carId },
+      where: { id: validatedCarId.data },
     });
   } catch (error) {
     return {
